+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "remark-retext",
   "version": "5.0.0",
-  "description": "remark plugin to transform to retext",
+  "description": "remark plugin to support retext",
   "license": "MIT",
   "keywords": [
     "unified",
@@ -30,6 +30,7 @@
   "main": "index.js",
   "types": "index.d.ts",
   "files": [
+    "lib/",
     "index.d.ts",
     "index.js"
   ],
@@ -56,7 +57,7 @@
     "xo": "^0.45.0"
   },
   "scripts": {
-    "build": "rimraf \"*.d.ts\" && tsc && type-coverage",
+    "build": "rimraf \"lib/**/*.d.ts\" \"*.d.ts\" && tsc && type-coverage",
     "format": "remark . -qfo && prettier . -w --loglevel warn && xo --fix",
     "test-api": "node --conditions development test.js",
     "test-coverage": "c8 --check-coverage --branches 100 --functions 100 --lines 100 --statements 100 --reporter lcov npm run test-api",
@@ -85,8 +86,8 @@
     "ignoreCatch": true,
     "#": "needed `any`s",
     "ignoreFiles": [
-      "index.d.ts",
-      "index.js"
+      "lib/index.d.ts",
+      "lib/index.js"
     ]
   }
 }
