+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["*.js"],
+  "include": ["lib/**/*.js", "*.js"],
   "compilerOptions": {
     "target": "ES2020",
     "lib": ["ES2020"],
